# docs/agents/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "drive-coding HTTP API",
    "version": "1.3.0",
    "description": "Documentable Hono routes on the drive-coding backend. Generated from code — run `bun scripts/gen-openapi.mjs` to refresh."
  },
  "paths": {
...
    "/api/agents/:id/history": {
      "get": {
        "operationId": "get_api_agents_id_history",
        "summary": "GET /api/agents/:id/history",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}